Charts > Multiple updates to horizontal bars
Moves horizontal label additions from LinearChart to BarChart.
This is done so that labels will display even when the axis is hidden in the future.

Adds checkbox for adding numeric values to bars, and removes the On Bar selection for now.
At the moment On Bar label types should be removed until we figure out a better way to manage the height state.

Adds minor helper utitlies for determing which type of labels to use on horizontal bars.

Uses Chroma to set label color