[DEV-9669] Handle Scatter Plot Continuous useScales
On scatter plots with continuous scales, the date/category axis didn't
have the ability to change the min/max of the scale. The resulted in large
amounts of padding being added between zero and the first value.

Adds TS improvements to useScale and useMinMax hooks.

Removes Scatterplot example to favor storybook examples.